Reorganized code; top level is cmd only.
