+++ radis/albumfolder_test.go
@@ -1,4 +1,4 @@
-package main
+package radis
 
 import (
 	"errors"
